rename script.js
